[MIG] medical_diagnostic_report: Finish migration
