+++ medical_diagnostic_report/templates/assets.xml
@@ -8,7 +8,7 @@
         <xpath expr="." position="inside">
             <script
                 type="text/javascript"
-                src="/medical_diagnostic_report/static/src/js/observation_value_widget.js"
+                src="/medical_diagnostic_report/static/src/js/list_renderer.js"
             />
             <script
                 type="text/javascript"
