+++ medical_diagnostic_report/views/medical_diagnostic_report.xml
@@ -17,7 +17,7 @@
                     class="btn-success"
                     confirm="Are you sure you want to validate this report?"
                     type="object"
-                    states="registered"
+                    attrs="{'invisible':[('fhir_state', '!=', 'registered')]}"
                 />
                 <button
                     name="cancel_action"
@@ -33,7 +33,7 @@
                     name="web_ribbon"
                     text="Cancelled"
                     bg_color="bg-danger"
-                    attrs="{'invisible':[('state', '!=', 'cancelled')]}"
+                    attrs="{'invisible':[('fhir_state', '!=', 'cancelled')]}"
                 />
             </xpath>
             <xpath expr="//div[@name='button_box']" position="inside">
@@ -75,9 +75,18 @@
             </xpath>
             <field name="occurrence_date" position="after">
                 <field name="issued_date" />
-                <field name="issued_user_id" states="final,cancelled" />
-                <field name="cancel_date" states="cancelled" />
-                <field name="cancel_user_id" states="cancelled" />
+                <field
+                    name="issued_user_id"
+                    attrs="{'invisible':[('fhir_state', 'not in', ['final', 'cancelled'])]}"
+                />
+                <field
+                    name="cancel_date"
+                    attrs="{'invisible':[('fhir_state', '!=', 'cancelled')]}"
+                />
+                <field
+                    name="cancel_user_id"
+                    attrs="{'invisible':[('fhir_state', '!=', 'cancelled')]}"
+                />
                 <field name="item_blocked" invisible="1" />
                 <field name="with_observation" invisible="1" />
                 <field name="with_conclusion" invisible="1" />
@@ -137,30 +146,29 @@
                                 widget="section_and_note_text"
                                 attrs="{'readonly': ['|', ('blocked', '=', True), ('concept_id', '!=', False)]}"
                             />
-                            <field name="value" invisible="1" />
+                            <field name="value" widget="diagnostic_report_value">
+
                             <!-- TODO: Fix this, it needs to be grouped somehow -->
-                            <field
-                                name="value_bool"
-                                attrs="{'invisible': [('value_type', '!=', 'bool')]}"
-                            />
-                            <field
-                                name="value_str"
-                                attrs="{'invisible': [('value_type', '!=', 'str')]}"
-                            />
-                            <field
+                                <field name="value_bool" />
+                                <field name="value_str" />
+                                <field
+                                    name="value_selection"
+                                    widget="dynamic_selection_diagnostic_report"
+                                    context="{'options': selection_options}"
+                                />
+                                <field name="value_int" />
+                                <field name="value_float" />
+                            </field>
+                            <field name="value_bool" invisible="1" />
+                                <field name="value_str" invisible="1" />
+                                <field
                                 name="value_selection"
                                 widget="dynamic_selection_diagnostic_report"
                                 context="{'options': selection_options}"
-                                attrs="{'invisible': [('value_type', '!=', 'selection')]}"
+                                invisible="1"
                             />
-                            <field
-                                name="value_int"
-                                attrs="{'invisible': [('value_type', '!=', 'int')]}"
-                            />
-                            <field
-                                name="value_float"
-                                attrs="{'invisible': [('value_type', '!=', 'float')]}"
-                            />
+                                <field name="value_int" invisible="1" />
+                                <field name="value_float" invisible="1" />
                             <!-- TODO: Unitl here....-->
                             <field
                                 name="uom_id"
